git-rev-list --objects gone
it was the wrong tool for the job.  i got hung up on a naively liberal
interpretation of "any object referenced by the listed commits" where
i should've been paying attention to "send me all object IDs which
I need to download if I have the commit object bar but not foo".
shame on me.

shame on git for not providing the raw-diff output in a sensible way
for a range of commits.  sensible excludes git-log, that's porcelain
in its own right.